[receiver/telemetryapi] add telemetry api receiver
This receiver listens for events from the Telemetry API and produces telemetry accordingly. The initial implementation will record a span for platform initialization events. Additional improvements would be to generate metrics for all reporting events.

Note that this span currently does not have a trace/span ID. The idea is to use a processor to correlate the event w/ the function that executed.

Signed-off-by: Alex Boten <aboten@lightstep.com>